PHP-DS V-1.0.0
You code a class with PHP, and you want your class to be able to translate to multiple languages. As you know, the arrangements made within the file are very troublesome. PHP "PHP-DS", with a different method, allows your classes to have an infrastructure that can support multiple languages.

This is a fully unique i18n experience!